trabajo_practico/AtaquesDePresentacion/requirements.txt
numpy>=1.19.3
opencv-python==4.4.0.46

Keras>=2.4.3
matplotlib>=3.3.3
tensorflow-gpu>=2.4.0
scikit-learn>=0.24.0
pandas>=1.2.0
scipy>=1.5.4